requirements.txt
matplotlib<=3.3.0,>=3.0.0
dill>=0.3.2
statsmodels>=0.11.1
Sphinx>=1.4.8
sphinx-rtd-theme>=0.4.3
sphinx-argparse>=0.2.5
docutils>=0.14.0
numpydoc>=1.1.0
requests>=2.21
joblib>=0.17.0
pandas>=0.24.2
psutil>=5.6.1
scipy>=1.4.1
setuptools>=47.3.0
nibabel>=3.1.1
indexed_gzip>=1.4.0
seaborn>=0.10.0
networkx==2.4
nipype>=1.5.1
scikit_learn>=0.23.0
dipy>=1.1.0
boto3>=1.9.111
pybids>=0.6.4
colorama>=0.3.9
python-louvain>=0.13
graspologic>=0.2.0
sqlalchemy~=1.3.1
#duecredit>=0.8.0
h5py>=2.10.0
scikit-image>=0.14.2
six>=1.12.0
pyyaml>=6.0
git+https://github.com/dPys/nilearn.git@enh/parc_conn
git+https://github.com/dPys/deepbrain.git@master
urllib3>=1.25.4
mplcyberpunk>=0.1.11
numba>=0.51.2
#asyncio>=3.4.3
#aiofiles>=0.8.0
